+++ kafka-balance-stream/src/main/kotlin/kafka/balance/stream/BalanceStream.kt
@@ -1,5 +1,6 @@
 package kafka.balance.stream
 
+import kafka.balance.stream.config.StreamConfig
 import kafka.balance.stream.serialization.BlockDeserializer
 import kafka.balance.stream.processor.AddressBalanceProcessorSupplier
 import org.apache.kafka.common.serialization.StringDeserializer
