+++ kafka-balance-stream/src/main/kotlin/kafka/balance/stream/config/StreamConfig.kt
@@ -1,28 +1,27 @@
-package kafka.lastSeen.stream
+package kafka.balance.stream.config
 
-import kafka.lastSeen.stream.serialization.AddressFeatureSerdes
+import kafka.balance.stream.serialization.AddressFeatureSerdes
 import org.apache.kafka.common.serialization.Serdes
 import org.apache.kafka.streams.StreamsConfig
 import org.apache.kafka.streams.state.KeyValueStore
 import org.apache.kafka.streams.state.StoreBuilder
-import org.apache.kafka.streams.state.StoreSupplier
 import org.apache.kafka.streams.state.Stores
 import java.util.*
 
 class StreamConfig {
     companion object {
-        fun getLasSeenStreamProperties(): Properties{
+        fun getStreamProperties(): Properties {
             return Properties().apply {
                 setProperty(StreamsConfig.BOOTSTRAP_SERVERS_CONFIG, "127.0.0.1:9092")
-                setProperty(StreamsConfig.APPLICATION_ID_CONFIG, "lastSeen")
+                setProperty(StreamsConfig.APPLICATION_ID_CONFIG, "balance")
                 setProperty(StreamsConfig.DEFAULT_KEY_SERDE_CLASS_CONFIG, Serdes.String().javaClass.name)
                 setProperty(StreamsConfig.DEFAULT_VALUE_SERDE_CLASS_CONFIG, AddressFeatureSerdes().javaClass.name)
             }
         }
 
-        fun getLastSeenStoreSupplier(): StoreBuilder<KeyValueStore<String, String>>{
+        fun getAddressBalanceStoreSupplier(): StoreBuilder<KeyValueStore<String, String>> {
             return Stores.keyValueStoreBuilder(
-                    Stores.persistentKeyValueStore("LastSeen"),
+                    Stores.persistentKeyValueStore("AddressBalance"),
                     Serdes.String(),
                     Serdes.String()
             ).withLoggingDisabled() //enable in production
