+++ kafka-lastSeen-stream/src/main/kotlin/kafka/lastSeen/stream/config/StreamConfig.kt
@@ -1,6 +1,6 @@
-package kafka.balance.stream
+package kafka.lastSeen.stream.config
 
-import kafka.balance.stream.serialization.AddressFeatureSerdes
+import kafka.lastSeen.stream.serialization.AddressFeatureSerdes
 import org.apache.kafka.common.serialization.Serdes
 import org.apache.kafka.streams.StreamsConfig
 import org.apache.kafka.streams.state.KeyValueStore
@@ -10,18 +10,18 @@ import java.util.*
 
 class StreamConfig {
     companion object {
-        fun getStreamProperties(): Properties {
+        fun getLasSeenStreamProperties(): Properties{
             return Properties().apply {
                 setProperty(StreamsConfig.BOOTSTRAP_SERVERS_CONFIG, "127.0.0.1:9092")
-                setProperty(StreamsConfig.APPLICATION_ID_CONFIG, "balance")
+                setProperty(StreamsConfig.APPLICATION_ID_CONFIG, "lastSeen")
                 setProperty(StreamsConfig.DEFAULT_KEY_SERDE_CLASS_CONFIG, Serdes.String().javaClass.name)
                 setProperty(StreamsConfig.DEFAULT_VALUE_SERDE_CLASS_CONFIG, AddressFeatureSerdes().javaClass.name)
             }
         }
 
-        fun getAddressBalanceStoreSupplier(): StoreBuilder<KeyValueStore<String, String>> {
+        fun getLastSeenStoreSupplier(): StoreBuilder<KeyValueStore<String, String>>{
             return Stores.keyValueStoreBuilder(
-                    Stores.persistentKeyValueStore("AddressBalance"),
+                    Stores.persistentKeyValueStore("LastSeen"),
                     Serdes.String(),
                     Serdes.String()
             ).withLoggingDisabled() //enable in production
